feat (operators) : Separate pipes and sources
